V5 improvements to VariantContext.  Now fully supports genotypes.  Filtering enabled.  Significant tests throughout system.  Support for rebuilding variant contexts from subsets of genotypes.  Some code cleanup around repository
git-svn-id: file:///humgen/gsa-scr1/gsa-engineering/svn_contents/trunk@2721 348d0f76-0448-11de-a6fe-93d51630548a

+++ java/src/org/broadinstitute/sting/playground/gatk/walkers/diagnostics/newvarianteval/SubstitutionEvaluation.java
@@ -44,9 +44,9 @@ public class SubstitutionEvaluation implements VariantEvaluation {
             for (char abase : BaseUtils.BASES) {
                 String altAllele = String.format("%c", abase);
 
-                if (BaseUtils.isTransition(rbase, abase)) {
+                if (BaseUtils.isTransition((byte)rbase, (byte)abase)) {
                     transitions += substitutions.get(refAllele, altAllele);
-                } else if (BaseUtils.isTransversion(rbase, abase)) {
+                } else if (BaseUtils.isTransversion((byte)rbase, (byte)abase)) {
                     transversions += substitutions.get(refAllele, altAllele);
                 } else {
                     unknown += substitutions.get(refAllele, altAllele);
